move package questionResponse to main package
