+++ src/main/java/com/codecool/codecoolquiz/questionResponse/QuestionResponseService.java
@@ -1,4 +1,4 @@
-package com.codecool.codecoolquiz.quiz.questionResponse;
+package com.codecool.codecoolquiz.questionResponse;
 
 import java.util.List;
