+++ src/main/java/com/codecool/codecoolquiz/quiz/QuizService.java
@@ -1,7 +1,7 @@
 package com.codecool.codecoolquiz.quiz;
 
 import com.codecool.codecoolquiz.category.Category;
-import com.codecool.codecoolquiz.quiz.questionResponse.QuestionResponse;
+import com.codecool.codecoolquiz.questionResponse.QuestionResponse;
 
 
 import java.util.List;
